# swagger/output.json
{
  "swagger": "2.0",
  "info": {
    "title": "GSES2 BTC application",
    "version": "1.0.0",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "tags": [
    {
      "name": "rate",
      "description": "Отримання поточного курсу BTC до UAH"
    },
    {
      "name": "subscription",
      "description": "Робота з підпискою"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/rate": {
      "get": {
        "tags": [
          "rate"
        ],
        "summary": "Отримати поточний курс BTC до UAH",
        "description": "Запит має повертати поточний курс BTC до UAH використовуючи будь-який third party сервіс з публічним АРІ",
        "operationId": "rate",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Повертається актуальний курс BTC до UAH",
            "schema": {
              "$ref": "#/definitions/integer"
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/subscribe": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Підписати емейл на отримання поточного курсу",
        "description": "Запит має перевірити, чи немає данної електронної адреси в поточній базі даних (файловій) і, в разі її відсутності, записувати її. Пізніше, за допомогою іншого запиту ми будемо відправляти лист на ті електронні адреси, які будуть в цій базі.",
        "operationId": "subscribe",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "Електронна адреса, яку потрібно підписати",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "E-mail додано"
          },
          "409": {
            "description": "Повертати, якщо e-mail вже є в базі даних (файловій)"
          }
        }
      }
    },
    "/sendEmails": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Відправити e-mail з поточним курсом на всі підписані електронні пошти.",
        "description": "Запит має отримувати актуальний курс BTC до UAH за допомогою third-party сервісу та відправляти його на всі електронні адреси, які були підписані раніше.",
        "operationId": "sendEmails",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "E-mailʼи відправлено."
          }
        }
      }
    }
  },
  "definitions": {
    "integer": {
      "type": "number",
      "example": 0
    }
  }
}